+++ docs/News.rst
@@ -8,6 +8,12 @@ News
 SQLObject (master)
 ==================
 
+Minor features
+--------------
+
+* Use ``module_loader.exec_module(module_loader.create_module())``
+  instead of ``module_loader.load_module()`` when available.
+
 Tests, CI
 ---------